pyproject.toml
[project]
name = "ariadne-agentic-rag"
version = "0.1.0"
description = "Event-Driven Agentic RAG for DIT UOP"
readme = "README.md"
license = "MIT"
authors = [
    {name="Georgios Giannopoulos", email="georgegiannopoulos2@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.5",
    "beautifulsoup4>=4.14.3",
    "chainlit>=2.11.1",
    "fake-useragent>=2.2.0",
    "fastembed>=0.8.0",
    "google-genai>=1.73.1",
    "llama-index-core>=0.14.21",
    "llama-index-embeddings-google-genai>=0.5.0",
    "llama-index-llms-google-genai>=0.9.1",
    "llama-index-utils-workflow>=0.10.1",
    "llama-index-vector-stores-qdrant>=0.10.0",
    "llama-index-vector-stores-redis>=0.8.0",
    "markdownify>=1.2.2",
    "openinference-instrumentation-llama-index>=4.3.9",
    "opentelemetry-exporter-otlp>=1.41.1",
    "opentelemetry-sdk>=1.41.1",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pymupdf>=1.27.2.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "python-frontmatter>=1.1.0",
    "python-json-logger>=4.1.0",
    "pyyaml>=6.0.3",
    "qdrant-client>=1.17.1",
    "redis>=5.3.1",
    "redisvl>=0.4.1",
    "sentencepiece>=0.2.1",
    "tenacity>=9.1.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ariadne", "pipelines"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
